+++ types/dev/vm.d.ts
@@ -15,8 +15,6 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import type {FindTermsMatchType, FindTermsSortOrder, FindTermsVariantMode, FindTermsEmphaticSequencesMode, FindKanjiDictionary, FindTermDictionary} from '../ext/translation';
-
 export type PseudoChrome = {
     runtime: {
         getURL(path: string): string;
@@ -29,41 +27,4 @@ export type PseudoFetchResponse = {
     statusText: string;
     text(): Promise<string>;
     json(): Promise<unknown>;
-};
-
-export type OptionsPresetObject = {
-    [key: string]: OptionsPreset;
-};
-
-export type OptionsList = string | (string | OptionsPreset)[];
-
-export type OptionsPreset = FindKanjiOptionsPreset | FindTermsOptionsPreset;
-
-export type FindKanjiOptionsPreset = {
-    enabledDictionaryMap?: [key: string, value: FindKanjiDictionary][];
-    removeNonJapaneseCharacters?: boolean;
-};
-
-export type FindTermsOptionsPreset = {
-    matchType?: FindTermsMatchType;
-    deinflect?: boolean;
-    mainDictionary?: string;
-    sortFrequencyDictionary?: string | null;
-    sortFrequencyDictionaryOrder?: FindTermsSortOrder;
-    removeNonJapaneseCharacters?: boolean;
-    convertHalfWidthCharacters?: FindTermsVariantMode;
-    convertNumericCharacters?: FindTermsVariantMode;
-    convertAlphabeticCharacters?: FindTermsVariantMode;
-    convertHiraganaToKatakana?: FindTermsVariantMode;
-    convertKatakanaToHiragana?: FindTermsVariantMode;
-    collapseEmphaticSequences?: FindTermsEmphaticSequencesMode;
-    textReplacements?: (FindTermsTextReplacement[] | null)[];
-    enabledDictionaryMap?: [key: string, value: FindTermDictionary][];
-    excludeDictionaryDefinitions?: string[] | null;
-};
-
-export type FindTermsTextReplacement = {
-    pattern: string;
-    flags: string;
-    replacement: string;
 };
